Consolidate ILibraryManager test stub and cover critical-exception rethrow
- Extract the duplicated DispatchProxy-based ILibraryManager fake from
  TestSkipIntroController and TestMediaSegmentRefreshService into a shared
  EntrypointTestHelpers.CreateLibraryManager helper
- Add a test asserting RefreshAsync rethrows critical/system exceptions
  instead of swallowing them
